requirements.txt
asttokens==3.0.1
decorator==5.2.1
executing==2.2.1
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.15.0
networkx==3.6.1
numpy==2.4.1
packaging==26.0
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
plotly==6.5.2
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pyvis==0.3.2
PyYAML==6.0.3
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.3.0
